Move knowhere/utils to knowhere/knowhere/common
